[#1] Modularized the router
