+++ const/env.js
@@ -1,6 +1,3 @@
 export const PORT = process.env.PORT || 8080;
-export const BASE_API_ENDPOINT = '/api/v1';
-export const WIDGETS_ENDPOINT = '/widgets';
-export const DOCUMENTS_ENDPOINT = '/documents';
-export const USERS_ENDPOINT = '/users';
-export const FLOW_MANAGEMENT_ENDPOINT = '/flow-management';
+export const API_VERSION = process.env.API_VERSION || 'v1';
+export const BASE_API_ENDPOINT = `/api/${API_VERSION}`;
