+++ modules/documents/const.js
@@ -1,0 +1,1 @@
+export const DOCUMENTS_ENDPOINT = '/documents';
