+++ modules/flows/const.js
@@ -1,0 +1,1 @@
+export const FLOWS_ENDPOINT = '/flows';
