+++ modules/flows/router/index.js
@@ -1,16 +1,16 @@
 import express from 'express';
-import { FLOW_MANAGEMENT_ENDPOINT as endpoint } from '../../const/env';
+import { FLOWS_ENDPOINT as endpoint } from '../const';
 
-const flowManagementRouter = express.Router();
+const flowsRouter = express.Router();
 const resourceType = 'flow management';
 
-flowManagementRouter.get('/', (req, res) => {
+flowsRouter.get('/', (req, res) => {
 	res.send(req.__('Hello from {{ endpoint }}', {
 		endpoint
 	}));
 });
 
-flowManagementRouter.get('/:flowId', (req, res) => {
+flowsRouter.get('/:flowId', (req, res) => {
 	const { flowId: id } = req.params;
 
 	res.send(req.__('Getting {{ resourceType }} resource with id {{ id }}', {
@@ -19,11 +19,11 @@ flowManagementRouter.get('/:flowId', (req, res) => {
 	}));
 });
 
-flowManagementRouter.post('/create-flow', (req, res) => {
+flowsRouter.post('/create-flow', (req, res) => {
 	res.send(req.__('Creating new flow...'));
 });
 
-flowManagementRouter.put('/:flowId', (req, res) => {
+flowsRouter.put('/:flowId', (req, res) => {
 	const { flowId: id } = req.params;
 
 	res.send(req.__('Updating {{ resourceType }} resource with id {{ id }}', {
@@ -32,7 +32,7 @@ flowManagementRouter.put('/:flowId', (req, res) => {
 	}));
 });
 
-flowManagementRouter.delete('/:flowId', (req, res) => {
+flowsRouter.delete('/:flowId', (req, res) => {
 	const { flowId: id } = req.params;
 
 	res.send(req.__('Deleting {{ resourceType }} resource with id {{ id }}', {
@@ -41,4 +41,4 @@ flowManagementRouter.delete('/:flowId', (req, res) => {
 	}));
 });
 
-export default flowManagementRouter;
+export default flowsRouter;
