+++ modules/users/const.js
@@ -1,0 +1,1 @@
+export const USERS_ENDPOINT = '/users';
