+++ modules/widgets/const.js
@@ -1,0 +1,1 @@
+export const WIDGETS_ENDPOINT = '/widgets';
