+++ modules/widgets/router/index.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import {WIDGETS_ENDPOINT as endpoint} from '../../const/env';
+import {WIDGETS_ENDPOINT as endpoint} from '../const';
 
 const widgetsRouter = express.Router();
 const resourceType = 'widget';
