+++ router/index.js
@@ -1,13 +1,12 @@
 import express from 'express';
-import {
-	WIDGETS_ENDPOINT, 
-	DOCUMENTS_ENDPOINT,
-	USERS_ENDPOINT,
-  FLOW_MANAGEMENT_ENDPOINT } from '../const/env';
-import widgetsRouter from './widget/index';
-import documentsRouter from './document/index';
-import usersRouter from './user/index';
-import flowManagementRouter from './flow-management';
+import widgetsRouter from '../modules/widgets/router';
+import documentsRouter from '../modules/documents/router';
+import usersRouter from '../modules/users/router';
+import flowsRouter from '../modules/flows/router';
+import {WIDGETS_ENDPOINT} from "../modules/widgets/const";
+import {USERS_ENDPOINT} from "../modules/users/const";
+import {FLOWS_ENDPOINT} from "../modules/flows/const";
+import {DOCUMENTS_ENDPOINT} from "../modules/documents/const";
 
 const router = express.Router();
 
@@ -17,7 +16,7 @@ router.use(DOCUMENTS_ENDPOINT, documentsRouter);
 
 router.use(USERS_ENDPOINT, usersRouter);
 
-router.use(FLOW_MANAGEMENT_ENDPOINT, flowManagementRouter);
+router.use(FLOWS_ENDPOINT, flowsRouter);
 
 router.get('/', (req, res) =>
 	res.send(req.__('Hello world!'))
